docs: add quiz analytics reference
