+++ app/flashoffer/docs/flashoffer-react/reference/README.md
@@ -21,6 +21,8 @@
 - [MultipleChoiceAnswer](interfaces/MultipleChoiceAnswer.md)
 - [MultipleChoiceOption](interfaces/MultipleChoiceOption.md)
 - [MultipleChoiceQuizProps](interfaces/MultipleChoiceQuizProps.md)
+- [QuizAnalyticsOptions](interfaces/QuizAnalyticsOptions.md)
+- [QuizCompletionPayload](interfaces/QuizCompletionPayload.md)
 - [PreviewCardProps](interfaces/PreviewCardProps.md)
 - [SectionHeaderProps](interfaces/SectionHeaderProps.md)
 - [SectionProps](interfaces/SectionProps.md)
@@ -50,6 +52,7 @@
 - [Footer](functions/Footer.md)
 - [HeroBanner](functions/HeroBanner.md)
 - [InstagramEngagementSection](functions/InstagramEngagementSection.md)
+- [logQuizCompletion](functions/logQuizCompletion.md)
 - [MultipleChoiceQuiz](functions/MultipleChoiceQuiz.md)
 - [OutlineCtaButton](functions/OutlineCtaButton.md)
 - [PreviewCard](functions/PreviewCard.md)
